refactor - move classes to the package to organize the classes
