+++ app/src/main/java/com/chachaup/chachaup/ui/SignUp.java
@@ -1,4 +1,4 @@
-package com.chachaup.chachaup;
+package com.chachaup.chachaup.ui;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -8,8 +8,9 @@ import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.chachaup.chachaup.R;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
